pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shencha-agent"
version = "2.1.0"
description = "AI-Powered Autonomous Code Audit Agent with Multi-Expert Team (UI, Product, Aesthetics, Architecture, Logic Masters)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "X-Tavern Team", email = "dev@x-tavern.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["code-audit", "ai", "llm", "claude", "automation", "agent"]

dependencies = [
    "claude-agent-sdk>=0.1.0",
    "anyio>=4.0.0",
    "aiofiles>=23.2.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "PyGithub>=2.1.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
shencha = "src.cli:main"
shencha-agent = "src.cli:main"
sc = "src.cli:main"

[project.urls]
Homepage = "https://github.com/x-tavern/shencha-agent"
Documentation = "https://github.com/x-tavern/shencha-agent#readme"
Repository = "https://github.com/x-tavern/shencha-agent"
Issues = "https://github.com/x-tavern/shencha-agent/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
